Fix bugs for controller

+++ src/controllers/catalog.controller.js
@@ -4,16 +4,22 @@ const { info, error } = require('../helpers/Logger');
 // TODO: implement validation, auth req
 const findAllProducts = async (req, res, next) => {
   try {
-    await Product.find()
-      .then((products) => {
-        if (products) {
-          return res.json(products)
-            .status(200);
-        }
-        res.sendStatus(401);
-      }).catch((err) => {
-        throw err;
-      });
+    const auth = true;
+
+    if (auth) {
+      await Product.find()
+        .then((products) => {
+          if (products) {
+            return res.json(products)
+              .status(200);
+          }
+          res.sendStatus(404);
+        }).catch((err) => {
+          throw err;
+        });
+    } else {
+      res.sendStatus(401);
+    }
   } catch (err) {
     onError(res, err);
   }
@@ -23,23 +29,28 @@ const findAllProducts = async (req, res, next) => {
 const findProductById = async (req, res, next) => {
   try {
     const { barcode } = req.params;
+    const auth = true;
 
-    await Post.findOne({
-      barcode: barcode
-    }).then((product) => {
-      if (product) {
-        return res.json(product)
-          .status(200);
-      }
+    if (auth) {
+      await Product.findOne({
+        barcode: barcode
+      }).then((product) => {
+        if (product) {
+          return res.json(product)
+            .status(200);
+        }
+        res.sendStatus(404);
+      }).catch((err) => {
+        if (err.kind === 'ObjectId') {
+          return res.json({
+            message: 'Product not found with barcode: ' + barcode
+          }).status(404);
+        }
+        throw err;
+      });
+    } else {
       res.sendStatus(401);
-    }).catch((err) => {
-      if (err.kind === 'ObjectId') {
-        return res.json({
-          message: 'Product not found with barcode: ' + barcode
-        }).status(404);                
-      }
-      throw err;
-    });
+    }
   } catch (err) {
     onError(res, err);
   }
@@ -48,23 +59,49 @@ const findProductById = async (req, res, next) => {
 // TODO: implement validation, auth req
 const createProduct = async (req, res, next) => {
   try {
-    const product = new Product({
-      barcode: req.body.barcode,
-      name: req.body.name,
-      description: req.body.description,
-      img: req.body.img,
-      stock: req.body.stock,
-      supplier: req.body.supplier,
-      available: req.body.available
-    });
+    const auth = true;
 
-    await product.save()
-      .then((data) => {
-        res.json(data)
-          .status(200);
-      }).catch((err) => {
-        throw err;
+    if (auth) {
+      const product = new Product({
+        barcode: req.body.barcode,
+        name: req.body.name,
+        description: req.body.description,
+        img: req.body.img,
+        stock: req.body.stock,
+        price: req.body.price,
+        supplier: req.body.supplier,
+        available: req.body.available
       });
+  
+      let hasProduct = false;
+  
+      await Product.findOne({
+        barcode: req.body.barcode
+      }).then((product) => {
+        if (product) {
+          hasProduct = true;
+        }
+      }).catch((err) => {
+        if (err.kind !== 'ObjectId') {
+          hasProduct = true;
+          throw err;
+        }
+      });
+  
+      if (hasProduct) {
+        res.sendStatus(404);
+      } else {
+        await product.save()
+          .then((data) => {
+            res.json(data)
+              .status(200);
+          }).catch((err) => {
+            throw err;
+          });
+      }
+    } else {
+      res.sendStatus(401);
+    }
   } catch (err) {
     onError(res, err);
   }
@@ -74,33 +111,39 @@ const createProduct = async (req, res, next) => {
 const updateProduct = async (req, res, next) => {
   try {
     const { barcode } = req.params;
+    const auth = true;
 
-    await Product.findOneAndUpdate({
-      barcode: barcode
-    }, {
-      barcode: req.body.barcode,
-      name: req.body.name,
-      description: req.body.description,
-      img: req.body.img,
-      stock: req.body.stock,
-      supplier: req.body.supplier,
-      available: req.body.available
-    }, {
-      new: true
-    }).then((product) => {
-      if (product) {
-        return res.json(product)
-          .status(200);
-      }
+    if (auth) {
+      await Product.findOneAndUpdate({
+        barcode: barcode
+      }, {
+        barcode: req.body.barcode,
+        name: req.body.name,
+        description: req.body.description,
+        img: req.body.img,
+        stock: req.body.stock,
+        price: req.body.price,
+        supplier: req.body.supplier,
+        available: req.body.available
+      }, {
+        new: true
+      }).then((product) => {
+        if (product) {
+          return res.json(product)
+            .status(200);
+        }
+        res.sendStatus(404);
+      }).catch((err) => {
+        if (err.kind === 'ObjectId') {
+          return res.json({
+            message: 'Product not found with barcode: ' + barcode
+          }).status(404);                
+        }
+        throw err;
+      });
+    } else {
       res.sendStatus(401);
-    }).catch((err) => {
-      if (err.kind === 'ObjectId') {
-        return res.json({
-          message: 'Product not found with barcode: ' + barcode
-        }).status(404);                
-      }
-      throw err;
-    });
+    }
   } catch (err) {
     onError(res, err);
   }
@@ -110,22 +153,27 @@ const updateProduct = async (req, res, next) => {
 const deleteProduct = async (req, res, next) => {
   try {
     const { barcode } = req.params;
+    const auth = true;
 
-    await Product.findOneAndRemove({
-      barcode: barcode
-    }).then((product) => {
-      if (product) {
-        return res.sendStatus(204);
-      }
+    if (auth) {
+      await Product.findOneAndRemove({
+        barcode: barcode
+      }).then((product) => {
+        if (product) {
+          return res.sendStatus(204);
+        }
+        res.sendStatus(404);
+      }).catch((err) => {
+        if (err.kind === 'ObjectId' || err.name === 'NotFound') {
+          return res.json({
+              message: 'Product not found with barcode: ' + barcode
+          }).status(404);
+        }
+        throw err;
+      });
+    } else {
       res.sendStatus(401);
-    }).catch((err) => {
-      if (err.kind === 'ObjectId' || err.name === 'NotFound') {
-        return res.json({
-            message: 'Product not found with barcode: ' + barcode
-        }).status(404);
-      }
-      throw err;
-    });
+    }
   } catch (err) {
     onError(res, err);
   }
